+++ contracts/interface/IAccountBalance.sol
@@ -105,7 +105,7 @@ interface IAccountBalance {
     /// @return netQuoteBalance = quote.balance + totalQuoteInPools
     function getNetQuoteBalance(address trader) external view returns (int256);
 
-    function getPositionSize(address trader, address baseToken) external view returns (int256);
+    function getTotalPositionSize(address trader, address baseToken) external view returns (int256);
 
     function getTakerPositionSize(address trader, address baseToken) external view returns (int256);