+++ test/clearingHouse/ClearingHouse.accounting.xyk.test.ts
@@ -228,8 +228,8 @@ describe("ClearingHouse accounting verification in xyk pool", () => {
     }
 
     async function getTakerMakerPositionSizeDelta(): Promise<BigNumberish> {
-        const takerPosSize = await accountBalance.getPositionSize(taker.address, baseToken.address)
-        const makerPosSize = await accountBalance.getPositionSize(maker.address, baseToken.address)
+        const takerPosSize = await accountBalance.getTotalPositionSize(taker.address, baseToken.address)
+        const makerPosSize = await accountBalance.getTotalPositionSize(maker.address, baseToken.address)
         return takerPosSize.add(makerPosSize)
     }
 
@@ -268,7 +268,7 @@ describe("ClearingHouse accounting verification in xyk pool", () => {
             expect(await getTakerMakerPositionSizeDelta()).be.closeTo(BigNumber.from(0), 2)
 
             await takerCloseEth()
-            expect((await accountBalance.getPositionSize(maker.address, baseToken.address)).toString()).eq("0")
+            expect((await accountBalance.getTotalPositionSize(maker.address, baseToken.address)).toString()).eq("0")
 
             await makerRemoveLiquidity()
         })
@@ -278,7 +278,7 @@ describe("ClearingHouse accounting verification in xyk pool", () => {
             expect(await getTakerMakerPositionSizeDelta()).be.closeTo(BigNumber.from(0), 2)
 
             await takerCloseEth()
-            expect((await accountBalance.getPositionSize(maker.address, baseToken.address)).toString()).eq("0")
+            expect((await accountBalance.getTotalPositionSize(maker.address, baseToken.address)).toString()).eq("0")
 
             await makerRemoveLiquidity()
         })
@@ -288,7 +288,7 @@ describe("ClearingHouse accounting verification in xyk pool", () => {
             expect(await getTakerMakerPositionSizeDelta()).be.closeTo(BigNumber.from(0), 2)
 
             await takerCloseEth()
-            expect((await accountBalance.getPositionSize(maker.address, baseToken.address)).toString()).eq("0")
+            expect((await accountBalance.getTotalPositionSize(maker.address, baseToken.address)).toString()).eq("0")
 
             await makerRemoveLiquidity()
         })
@@ -298,7 +298,7 @@ describe("ClearingHouse accounting verification in xyk pool", () => {
             expect(await getTakerMakerPositionSizeDelta()).be.closeTo(BigNumber.from(0), 2)
 
             await takerCloseEth()
-            expect((await accountBalance.getPositionSize(maker.address, baseToken.address)).toString()).eq("0")
+            expect((await accountBalance.getTotalPositionSize(maker.address, baseToken.address)).toString()).eq("0")
 
             await makerRemoveLiquidity()
         })
@@ -397,7 +397,7 @@ describe("ClearingHouse accounting verification in xyk pool", () => {
 
             // taker close
             await closePosition(fixture, taker)
-            expect(await accountBalance.getPositionSize(maker.address, baseToken.address)).to.be.deep.eq(0)
+            expect(await accountBalance.getTotalPositionSize(maker.address, baseToken.address)).to.be.deep.eq(0)
 
             // maker account value = freeCollateral + unsettled PnL
             const makerAccountValue = await clearingHouse.getAccountValue(maker.address)
@@ -428,7 +428,7 @@ describe("ClearingHouse accounting verification in xyk pool", () => {
             await closePosition(fixture, taker2)
             await closePosition(fixture, taker3)
 
-            expect(await accountBalance.getPositionSize(maker.address, baseToken.address)).to.be.deep.eq(0)
+            expect(await accountBalance.getTotalPositionSize(maker.address, baseToken.address)).to.be.deep.eq(0)
 
             // maker remove liquidity
             await removeAllOrders(fixture, maker)
@@ -449,9 +449,9 @@ describe("ClearingHouse accounting verification in xyk pool", () => {
             await closePosition(fixture, taker2)
             await closePosition(fixture, taker3)
 
-            const maker1PositionSize = await accountBalance.getPositionSize(maker.address, baseToken.address)
-            const maker2PositionSize = await accountBalance.getPositionSize(maker2.address, baseToken.address)
-            const maker3PositionSize = await accountBalance.getPositionSize(maker3.address, baseToken.address)
+            const maker1PositionSize = await accountBalance.getTotalPositionSize(maker.address, baseToken.address)
+            const maker2PositionSize = await accountBalance.getTotalPositionSize(maker2.address, baseToken.address)
+            const maker3PositionSize = await accountBalance.getTotalPositionSize(maker3.address, baseToken.address)
             expect(maker1PositionSize.add(maker2PositionSize).add(maker3PositionSize)).to.be.closeTo("0", 10)
 
             // makers remove liquidity
@@ -475,8 +475,8 @@ describe("ClearingHouse accounting verification in xyk pool", () => {
             // taker close position
             await closePosition(fixture, taker)
 
-            const maker1PositionSize = await accountBalance.getPositionSize(maker.address, baseToken.address)
-            const maker2PositionSize = await accountBalance.getPositionSize(maker2.address, baseToken.address)
+            const maker1PositionSize = await accountBalance.getTotalPositionSize(maker.address, baseToken.address)
+            const maker2PositionSize = await accountBalance.getTotalPositionSize(maker2.address, baseToken.address)
             expect(maker1PositionSize.add(maker2PositionSize)).to.be.closeTo("0", 10)
 
             // maker remove liquidity
@@ -503,7 +503,7 @@ describe("ClearingHouse accounting verification in xyk pool", () => {
             // taker close
             await closePosition(fixture, taker)
 
-            expect(await accountBalance.getPositionSize(maker.address, baseToken.address)).to.be.deep.eq(0)
+            expect(await accountBalance.getTotalPositionSize(maker.address, baseToken.address)).to.be.deep.eq(0)
 
             // maker remove liquidity
             await removeAllOrders(fixture, maker)
@@ -524,7 +524,7 @@ describe("ClearingHouse accounting verification in xyk pool", () => {
             // taker close
             await closePosition(fixture, taker)
 
-            expect(await accountBalance.getPositionSize(maker.address, baseToken.address)).to.be.deep.eq(0)
+            expect(await accountBalance.getTotalPositionSize(maker.address, baseToken.address)).to.be.deep.eq(0)
 
             // maker remove liquidity
             await removeAllOrders(fixture, maker)
@@ -564,11 +564,11 @@ describe("ClearingHouse accounting verification in xyk pool", () => {
             })
 
             // liquidate taker
-            while ((await accountBalance.getPositionSize(taker.address, baseToken.address)).gt(0)) {
+            while ((await accountBalance.getTotalPositionSize(taker.address, baseToken.address)).gt(0)) {
                 await clearingHouse.connect(taker2).liquidate(taker.address, baseToken.address)
             }
 
-            expect(await accountBalance.getPositionSize(maker.address, baseToken.address)).to.be.deep.eq(0)
+            expect(await accountBalance.getTotalPositionSize(maker.address, baseToken.address)).to.be.deep.eq(0)
 
             // maker remove liquidity
             await removeAllOrders(fixture, maker)
@@ -593,7 +593,7 @@ describe("ClearingHouse accounting verification in xyk pool", () => {
             }
 
             // liquidate taker
-            while ((await accountBalance.getPositionSize(taker.address, baseToken.address)).gt(0)) {
+            while ((await accountBalance.getTotalPositionSize(taker.address, baseToken.address)).gt(0)) {
                 await clearingHouse.connect(taker2).liquidate(taker.address, baseToken.address)
             }
         })