+++ test/clearingHouse/ClearingHouse.funding.test.ts
@@ -1149,11 +1149,11 @@ describe("ClearingHouse funding", () => {
 //     "pendingFundingPayment: ",
 //     (await exchange.getPendingFundingPayment(carol.address, baseToken.address)).toString(),
 // )
-// console.log("positionSize: ", (await accountBalance.getPositionSize(carol.address, baseToken.address)).toString())
+// console.log("positionSize: ", (await accountBalance.getTotalPositionSize(carol.address, baseToken.address)).toString())
 // console.log("alice")
 // console.log(
 //     "pendingFundingPayment: ",
 //     (await exchange.getPendingFundingPayment(alice.address, baseToken.address)).toString(),
 // )
-// console.log("positionSize: ", (await accountBalance.getPositionSize(alice.address, baseToken.address)).toString())
+// console.log("positionSize: ", (await accountBalance.getTotalPositionSize(alice.address, baseToken.address)).toString())
 // // === useful console.log for verifying stats ===