+++ test/clearingHouse/ClearingHouse.getPositionSize.test.ts
@@ -18,7 +18,7 @@ import { deposit } from "../helper/token"
 import { encodePriceSqrt } from "../shared/utilities"
 import { createClearingHouseFixture } from "./fixtures"
 
-describe("ClearingHouse.getPositionSize", () => {
+describe("ClearingHouse.getTotalPositionSize", () => {
     const [admin, alice, bob, carol] = waffle.provider.getWallets()
     const loadFixture: ReturnType<typeof waffle.createFixtureLoader> = waffle.createFixtureLoader([admin])
     let clearingHouse: TestClearingHouse
@@ -70,7 +70,7 @@ describe("ClearingHouse.getPositionSize", () => {
         })
 
         it("size = 0, if no swap", async () => {
-            expect(await accountBalance.getPositionSize(alice.address, baseToken.address)).eq(0)
+            expect(await accountBalance.getTotalPositionSize(alice.address, baseToken.address)).eq(0)
 
             await clearingHouse.connect(alice).addLiquidity({
                 baseToken: baseToken.address,
@@ -84,7 +84,7 @@ describe("ClearingHouse.getPositionSize", () => {
                 deadline: ethers.constants.MaxUint256,
             })
 
-            expect(await accountBalance.getPositionSize(alice.address, baseToken.address)).eq(0)
+            expect(await accountBalance.getTotalPositionSize(alice.address, baseToken.address)).eq(0)
         })
 
         it("bob(taker) swaps 1 time", async () => {
@@ -119,12 +119,14 @@ describe("ClearingHouse.getPositionSize", () => {
 
             // The swap pushes the mark price to 149.863446 (tick = 50099.75001)
 
-            expect(await accountBalance.getPositionSize(alice.address, baseToken.address)).eq(
+            expect(await accountBalance.getTotalPositionSize(alice.address, baseToken.address)).eq(
                 parseEther("0.408410420499999999"),
             )
 
             // 0.4084104205
-            expect(await accountBalance.getPositionSize(bob.address, baseToken.address)).eq(parseEther("-0.4084104205"))
+            expect(await accountBalance.getTotalPositionSize(bob.address, baseToken.address)).eq(
+                parseEther("-0.4084104205"),
+            )
         })
 
         it("bob swaps 2 time", async () => {
@@ -171,12 +173,14 @@ describe("ClearingHouse.getPositionSize", () => {
 
             // which makes the mark price become 149.863446 (tick = 50099.75001)
 
-            expect(await accountBalance.getPositionSize(alice.address, baseToken.address)).eq(
+            expect(await accountBalance.getTotalPositionSize(alice.address, baseToken.address)).eq(
                 parseEther("0.408410420599999999"),
             )
 
             // 0.2042052103 * 2 = 0.4084104206
-            expect(await accountBalance.getPositionSize(bob.address, baseToken.address)).eq(parseEther("-0.4084104206"))
+            expect(await accountBalance.getTotalPositionSize(bob.address, baseToken.address)).eq(
+                parseEther("-0.4084104206"),
+            )
         })
     })
 
@@ -255,16 +259,18 @@ describe("ClearingHouse.getPositionSize", () => {
         // mark price should be 153.8170921 (tick = 50360.15967)
 
         // -(1.633641682 / 2 + 0.6482449586) = -1.4650657996
-        expect(await accountBalance.getPositionSize(alice.address, baseToken.address)).eq(
+        expect(await accountBalance.getTotalPositionSize(alice.address, baseToken.address)).eq(
             parseEther("-1.465065799750044640"),
         )
 
         // 1.633641682 + 0.6482449586 = 2.2818866406
-        expect(await accountBalance.getPositionSize(bob.address, baseToken.address)).eq(
+        expect(await accountBalance.getTotalPositionSize(bob.address, baseToken.address)).eq(
             parseEther("2.281886640750044638"),
         )
 
         // -1.633641682 / 2 = -0.816820841
-        expect(await accountBalance.getPositionSize(carol.address, baseToken.address)).eq(parseEther("-0.816820841"))
+        expect(await accountBalance.getTotalPositionSize(carol.address, baseToken.address)).eq(
+            parseEther("-0.816820841"),
+        )
     })
 })